refactor(tunnel): lift local-forward engine into core
Move the local (ssh -L) forwarder, its channel-opener seam, and the
LocalForwardConfig/TunnelStats data types from the desktop tunnel module
into termihub_core::tunnel so the identical engine can run on the desktop
or on a remote agent (S3, part of #2139). The desktop tunnel module
re-exports them, so every existing call site and desktop behaviour is
unchanged. Adds a ReachableFrom classifier for agent-hosted listen
sockets per the endpoint-semantics concept.

Part of #2185
